# api/src/tsoa/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Achievement": {
				"type": "string"
			},
			"Achievements": {
				"properties": {
					"contractAddress": {
						"type": "string"
					},
					"nftId": {
						"type": "string"
					},
					"achievements": {
						"items": {
							"$ref": "#/components/schemas/Achievement"
						},
						"type": "array"
					}
				},
				"required": [
					"contractAddress",
					"nftId",
					"achievements"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Error": {
				"properties": {
					"error": {
						"type": "string",
						"enum": [
							"Achievements not found"
						],
						"nullable": false
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Response": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/Achievements"
					},
					{
						"$ref": "#/components/schemas/Error"
					}
				]
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@game/api",
		"version": "1.0.0",
		"description": "Webserver REST API for the middleware to interact with game node.",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/achievements": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "wallet_address",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/submissions_data": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "wallet_address",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/my_test": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/user_submission": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "wallet_address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "submission_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}